+++ src/cacheLRU_/NodoLista.java
@@ -1,4 +1,7 @@
 package cacheLRU_;
+
+import java.math.BigInteger;
+
 /**
  * @author Javier Toro Toro
  * @author Sebastián Cádiz Quezada
@@ -8,9 +11,9 @@ public class NodoLista {
     NodoLista next;
     NodoLista prev;
     int clave;
-    int resultado;
+    BigInteger resultado;
 
-    public NodoLista(int clave, int resultado) {
+    public NodoLista(int clave, BigInteger resultado) {
         this.next = null;
         this.prev = null;
         this.clave = clave;
